User find name

See merge request Krat12/jmstackoverflow!65

+++ src/main/java/com/javamentor/qa/platform/models/dto/UserDtoList.java
@@ -15,10 +15,10 @@ public class UserDtoList {
     private Long id;
     private String fullName;
     private String linkImage;
-    private Long reputation;
+    private Integer reputation;
     private List<TagDto> tags;
 
-    public UserDtoList(Long id, String fullName, String linkImage, Long reputation) {
+    public UserDtoList(Long id, String fullName, String linkImage, Integer reputation) {
         this.id = id;
         this.fullName = fullName;
         this.linkImage = linkImage;
